Do not test/build vdrtools anoncreds in CI, change aries-vcx default features
