+++ .github/workflows/main.yml
@@ -118,7 +118,7 @@ jobs:
     runs-on: ubuntu-20.04
     strategy:
       matrix:
-        backend: ["vdrtools", "modular_libs", "vdr_proxy_ledger"]
+        backend: ["modular_libs", "vdr_proxy_ledger"]
     steps:
       - name: "Git checkout"
         uses: actions/checkout@v3
@@ -141,7 +141,7 @@ jobs:
     runs-on: ubuntu-20.04
     strategy:
       matrix:
-        backend: ["vdrtools_anoncreds", "modular_libs", "vdr_proxy_ledger"]
+        backend: ["modular_libs", "vdr_proxy_ledger"]
     steps:
       - name: "Git checkout"
         uses: actions/checkout@v3
@@ -162,9 +162,6 @@ jobs:
 
   libvcx_core_clippy:
     runs-on: ubuntu-20.04
-    strategy:
-      matrix:
-        backend: ["anoncreds_credx", "anoncreds_vdrtools"]
     steps:
       - name: "Git checkout"
         uses: actions/checkout@v3
@@ -179,15 +176,12 @@ jobs:
           sudo apt-get install -y libsodium-dev libssl-dev libzmq3-dev
       - name: "Verify clippy across the entire workspace with default features"
         run: |
-          cargo clippy -p libvcx_core --features ${{ matrix.backend }}
+          cargo clippy -p libvcx_core --features anoncreds_credx
         env:
           RUSTFLAGS: -D warnings
 
   node_wrapper_clippy:
     runs-on: ubuntu-20.04
-    strategy:
-      matrix:
-        backend: ["anoncreds_credx", "anoncreds_vdrtools"]
     steps:
       - name: "Git checkout"
         uses: actions/checkout@v3
@@ -202,7 +196,7 @@ jobs:
           sudo apt-get install -y libsodium-dev libssl-dev libzmq3-dev
       - name: "Verify clippy across the entire workspace with default features"
         run: |
-          cargo clippy -p vcx-napi-rs --features ${{ matrix.backend }}
+          cargo clippy -p vcx-napi-rs --features anoncreds_credx
         env:
           RUSTFLAGS: -D warnings
 
@@ -306,12 +300,6 @@ jobs:
   test-unit-workspace:
     needs: workflow-setup
     runs-on: ubuntu-20.04
-    strategy:
-      matrix:
-        features: [
-          "anoncreds_vdrtools",
-          "anoncreds_credx",
-        ]
     steps:
       - name: "Git checkout"
         uses: actions/checkout@v3
@@ -321,17 +309,11 @@ jobs:
           rust-toolchain-version: ${{ env.RUST_TOOLCHAIN_VERSON }}
           skip-docker-setup: true
       - name: "Run workspace unit tests"
-        run: RUST_TEST_THREADS=1 cargo test --workspace --lib --exclude aries-vcx-agent --exclude libvdrtools --exclude wallet_migrator --features ${{ matrix.features }}
+        run: RUST_TEST_THREADS=1 cargo test --workspace --lib --exclude aries-vcx-agent --exclude libvdrtools --exclude wallet_migrator --features anoncreds_credx
 
   test-integration-aries-vcx:
     needs: workflow-setup
     runs-on: ubuntu-20.04
-    strategy:
-      matrix:
-        features: [
-          "vdrtools",
-          "modular_libs"
-        ]
     steps:
       - name: "Git checkout"
         uses: actions/checkout@v3
@@ -340,7 +322,7 @@ jobs:
         with:
           rust-toolchain-version: ${{ env.RUST_TOOLCHAIN_VERSON }}
       - name: "Run aries-vcx integration tests"
-        run: RUST_TEST_THREADS=1 cargo test --manifest-path="aries_vcx/Cargo.toml" --features ${{ matrix.features }} -- --ignored;
+        run: RUST_TEST_THREADS=1 cargo test --manifest-path="aries_vcx/Cargo.toml" --features modular_libs -- --ignored;
 
   test-integration-aries-vcx-mysql:
     needs: workflow-setup
@@ -406,12 +388,6 @@ jobs:
     needs: workflow-setup
     if: ${{ needs.workflow-setup.outputs.SKIP_CI != 'true' }}
     runs-on: ubuntu-20.04
-    strategy:
-      matrix:
-        features: [
-          "anoncreds_vdrtools",
-          "anoncreds_credx"
-        ]
     steps:
       - name: "Git checkout"
         uses: actions/checkout@v3
@@ -421,7 +397,7 @@ jobs:
           rust-toolchain-version: ${{ env.RUST_TOOLCHAIN_VERSON }}
       - name: "Run libvcx_core integration tests"
         run: |
-          RUST_TEST_THREADS=1 cargo test --features ${{ matrix.features }} --manifest-path="libvcx_core/Cargo.toml" -- --include-ignored;
+          RUST_TEST_THREADS=1 cargo test --features anoncreds_credx --manifest-path="libvcx_core/Cargo.toml" -- --include-ignored;
 
   test-integration-did-crate:
     needs: workflow-setup
