+++ aries_vcx/Cargo.toml
@@ -11,13 +11,13 @@ path = "src/lib.rs"
 doctest = false
 
 [features]
-default = ["vdrtools"]
+default = ["modular_libs", "aries_vcx_core/vdrtools_wallet"]
 vdrtools = ["aries_vcx_core/vdrtools_anoncreds", "aries_vcx_core/vdrtools_wallet"]
 # Feature flag to include the 'modular library' dependencies (vdrtools alternatives; indy-vdr, indy-credx)
 modular_libs = ["aries_vcx_core/modular_libs"]
 # TODO: Remove using "vdrtools" feature flag for vdr_proxy_ledger once IndyCredxAnonCreds
 # is fully implemented
-vdr_proxy_ledger = ["aries_vcx_core/vdr_proxy_ledger", "vdrtools"]
+vdr_proxy_ledger = ["aries_vcx_core/vdr_proxy_ledger"]
 
 # Feature for allowing legacy proof verification
 legacy_proof = ["aries_vcx_core/legacy_proof"]
