+++ aries_vcx/src/common/mod.rs
@@ -5,6 +5,4 @@ pub mod ledger;
 pub mod primitives;
 pub mod proofs;
 pub mod signing;
-#[cfg(feature = "vdrtools")]
-// TODO: Used by tests/ so not "hideable" by #[cfg(test)]
 pub mod test_utils;
