+++ aries_vcx/src/utils/mod.rs
@@ -8,7 +8,6 @@ use crate::{
 };
 
 #[macro_use]
-#[cfg(feature = "vdrtools")]
 pub mod devsetup;
 
 #[cfg(debug_assertions)]
@@ -32,8 +31,6 @@ pub mod constants;
 pub mod file;
 pub mod mockdata;
 pub mod openssl;
-#[cfg(feature = "vdrtools")]
-pub mod provision;
 pub mod qualifier;
 pub mod random;
 pub mod uuid;
