+++ uniffi_aries_vcx/core/Cargo.toml
@@ -17,7 +17,7 @@ path = "uniffi-bindgen.rs"
 
 [dependencies]
 uniffi = "0.23.0"
-aries-vcx = { path = "../../aries_vcx" }
+aries-vcx = { path = "../../aries_vcx", features = ["vdrtools"] }
 tokio = { version = "1.24.1" }
 once_cell = "1.17.0"
 thiserror = "1.0.38"
